Create lambda_function.py
An Amazon S3 trigger that retrieves metadata for the object that has been updated. The email is sent to the mail address mentioned in the S3 file and SMS notification is sent to the phone number in the S3 file with the message provided in the file stored in S3 bucket.